Restructure pages & components by moving/grouping as necessary
